requirements.txt
flask==3.0.0
flask-cors==4.0.0
google-cloud-firestore==2.16.0
pyyaml==6.0.1
python-dotenv==1.0.0
pyjwt==2.8.0
cryptography==41.0.7
requests==2.31.0
openai==1.12.0
anthropic==0.67.0
google-generativeai==0.8.0
gunicorn==21.2.0